requirements.txt
asgiref==3.2.10
astroid==2.4.2
Babel==2.8.0
Django==3.1.1
django-filter==2.3.0
django-phone-field==1.8.1
django-phonenumber-field==5.0.0
djangorestframework==3.11.1
gunicorn==20.0.4
isort==5.5.1
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
phonenumbers==8.12.9
psycopg2==2.8.5
pylint==2.6.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
wrapt==1.12.1
